Moved installation related functions to a new file. Replaced operators with functions
